# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Wishlist management service for creating and sharing wishlists",
        "title": "Wishlist API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/wishlists": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns all wishlists for authenticated user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wishlists"
                ],
                "summary": "Get user wishlists",
                "responses": {
                    "200": {
                        "description": "List of wishlists",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/domain.Wishlist"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a wishlist for authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wishlists"
                ],
                "summary": "Create new wishlist",
                "parameters": [
                    {
                        "description": "Wishlist data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.CreateWishlistRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Wishlist created",
                        "schema": {
                            "$ref": "#/definitions/responses.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/wishlists/delete": {
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes wishlist and all its gifts",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wishlists"
                ],
                "summary": "Delete wishlist",
                "parameters": [
                    {
                        "description": "Wishlist ID",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.DeleteWishlistRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Wishlist deleted",
                        "schema": {
                            "$ref": "#/definitions/responses.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Wishlist not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/wishlists/get": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns specific wishlist for authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wishlists"
                ],
                "summary": "Get wishlist by ID",
                "parameters": [
                    {
                        "description": "Wishlist ID",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.GetWishlistRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Wishlist found",
                        "schema": {
                            "$ref": "#/definitions/domain.Wishlist"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Wishlist not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/wishlists/public": {
            "get": {
                "description": "Public access to wishlist without authentication",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wishlists"
                ],
                "summary": "Get wishlist by share token",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Share token",
                        "name": "token",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Wishlist found",
                        "schema": {
                            "$ref": "#/definitions/domain.Wishlist"
                        }
                    },
                    "400": {
                        "description": "Token required",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Wishlist not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/wishlists/update": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates existing wishlist data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "wishlists"
                ],
                "summary": "Update wishlist",
                "parameters": [
                    {
                        "description": "Updated wishlist data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/requests.UpdateWishlistRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Wishlist updated",
                        "schema": {
                            "$ref": "#/definitions/responses.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Wishlist not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "domain.Gift": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "is_reserved": {
                    "type": "boolean"
                },
                "link": {
                    "type": "string"
                },
                "priority": {
                    "type": "integer"
                },
                "reserved_by": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "wishlist_id": {
                    "type": "integer"
                }
            }
        },
...
        "domain.Wishlist": {
            "type": "object",
            "properties": {
                "desc": {
                    "type": "string"
                },
                "event_date": {
                    "type": "string"
                },
                "gifts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.Gift"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "share_token": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
...
        "requests.CreateWishlistRequest": {
            "type": "object",
            "properties": {
                "desc": {
                    "type": "string"
                },
                "eventDate": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "requests.DeleteWishlistRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
        "requests.GetWishlistRequest": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                }
            }
        },
...
        "requests.UpdateWishlistRequest": {
            "type": "object",
            "properties": {
                "desc": {
                    "type": "string"
                },
                "eventDate": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "responses.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter JWT token in format: Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}